requirements.txt
colormath==3.0.0
Cython==0.29.6
decorator==4.4.0
networkx==2.3
numpy==1.16.2
Pillow==6.0.0
scipy==1.2.1
